1) switched from float to double data type 2) extracted battery info from pedelec entity into its own entity 3) made the necessary changes in other classes to handle the new entity relationship 4) pom.xml change: not ready for syslog yet

+++ src/main/java/de/rwth/idsg/bikeman/app/dto/ViewStationSlotsDTO.java
@@ -12,10 +12,10 @@ public class ViewStationSlotsDTO {
     private Integer stationSlotPosition;
     private OperationState state;
     private Boolean isOccupied;
-    private Float stateOfCharge;
+    private Double stateOfCharge;
 
     public ViewStationSlotsDTO(Long stationSlotId, Integer stationSlotPosition,
-                               OperationState state, Boolean isOccupied, Float stateOfCharge) {
+                               OperationState state, Boolean isOccupied, Double stateOfCharge) {
         this.stationSlotId = stationSlotId;
         this.stationSlotPosition = stationSlotPosition;
         this.state = state;
